+++ src/PAL/COM/sockets/ssl/mbedTLS/mbedtls.h
@@ -27,6 +27,11 @@ typedef struct mbedTLS_NFContext
     mbedtls_net_context* server_fd;
 }mbedTLS_NFContext;
 
+int net_would_block( const mbedtls_net_context *ctx );
+int mbedtls_net_recv( void *ctx, unsigned char *buf, size_t len );
+int mbedtls_net_send( void *ctx, const unsigned char *buf, size_t len );
+int mbedtls_net_recv_timeout( void *ctx, unsigned char *buf, size_t len, uint32_t timeout );
+
 // debug output declaration 
 void nf_debug( void *ctx, int level, const char *file, int line, const char *str );
